Added Baum Welch Trainiing

+++ HMMTransition.cpp
@@ -1,6 +1,21 @@
+/*
+ * HMMTransition.cpp
+ *
+ *	This is the cpp file for the HMMTransition object. HMMTransition
+ *  represents a transtion for a hidden markov model.  It simply consists
+ *  of a start node and end node.  State information from each node can be
+ *  used to determine what type of transition this is (same state to same
+ *  state? state1 to state2, state2 to state1, etc.).
+ *
+ *
+ *  Created on: 2-15-13
+ *      Author: tomkolar
+ */
 #include "HMMTransition.h"
 #include "HiddenMarkovModel.h"
 
+// Constuctors
+// ==============================================
 HMMTransition::HMMTransition() {
 }
 
@@ -10,10 +25,19 @@ HMMTransition::HMMTransition(HMMNode* aStartNode, HMMNode* anEndNode, HiddenMark
 	model = aModel;
 }
 
+// Destructor
+// =============================================
 HMMTransition::~HMMTransition() {
 }
 
-double HMMTransition::logProbability() {
+// Public Methods
+// =============================================
+
+// double logProbability()
+//  Purpose: 
+//		Returns the log of the transition probability for transitioning
+//		from the startNode to the endNode
+long double HMMTransition::logProbability() {
 	if (startNode->residue == HMMNode::startNodeChar)
 		return model->probabilities->logInitiationProbability(endNode->state);
 	else
